# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/EventEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Show.java
#	backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/basetest/TestData.java

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/ArtistsApi.java
@@ -111,7 +111,7 @@ public interface ArtistsApi {
         produces = { "application/json" }
     )
     default ResponseEntity<ArtistDto> artistsIdGet(
-        @Parameter(name = "id", description = "ID of the artist that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the artist that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         getRequest().ifPresent(request -> {
             for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
